policy Perfil

+++ app/Http/Controllers/PerfilController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Models\Perfil;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PerfilController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -14,19 +16,57 @@ class PerfilController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('index-perfil')) {
+            abort(403);
+        }
         return Perfil::all();
     }
 
 
     /**
-     * Display a listing of the resource to a front select.
+     * Display a listing of the resource with all relationships.
      *
      * @return \Illuminate\Http\Response
-    */
-    public function front_select()
+     */
+    public function index_full()
     {
-        return Perfil::all(['id as value','descricao as label']);
+        if (! Gate::allows('index-perfil')) {
+            abort(403);
+        }
+        return Perfil::all()->load('perfil');
     }
+
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Perfil  $perfil
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Perfil $perfil)
+    {
+        if (! Gate::allows('show-perfil', $perfil)) {
+            abort(403);
+        }
+        return $perfil;
+    }
+
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Perfil  $perfil
+     * @return \Illuminate\Http\Response
+     */
+    public function show_full(Perfil $perfil)
+    {
+        if (! Gate::allows('show-perfil', $perfil)) {
+            abort(403);
+        }
+        return $perfil->load('perfil');
+    }
+
+
 
     /**
      * Store a newly created resource in storage.
@@ -36,6 +76,9 @@ class PerfilController extends Controller
      */
     public function store(Request $request)
     {
+        if (! Gate::allows('create-perfil')) {
+            abort(403);
+        }
         $request->validate(
             rules: [
                 'descricao' => "required",
@@ -46,16 +89,6 @@ class PerfilController extends Controller
         );
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Perfil  $perfil
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Perfil $perfil)
-    {
-        return $perfil;
-    }
 
     /**
      * Update the specified resource in storage.
@@ -66,6 +99,9 @@ class PerfilController extends Controller
      */
     public function update(Request $request, Perfil $perfil)
     {
+        if (! Gate::allows('update-perfil', $perfil)) {
+            abort(403);
+        }
         $request->validate(
             rules: [
                 'descricao' => "required",
@@ -77,6 +113,7 @@ class PerfilController extends Controller
         return $perfil;
     }
 
+
     /**
      * Remove the specified resource from storage.
      *
@@ -85,6 +122,20 @@ class PerfilController extends Controller
      */
     public function destroy(Perfil $perfil)
     {
+        if (! Gate::allows('delete-perfil', $perfil)) {
+            abort(403);
+        }
         return $perfil->delete();
+    }
+
+
+    /**
+     * Display a listing of the resource to a front select.
+     *
+     * @return \Illuminate\Http\Response
+    */
+    public function front_select()
+    {
+        return Perfil::all(['id as value','descricao as text']);
     }
 }
